backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.2
torch>=1.9.0
opencv-python>=4.5.3
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
python-jose>=3.3.0
pydantic>=1.8.2
aiofiles>=0.7.0
python-dotenv>=0.19.0
requests>=2.31.0  # Added for testing
pytest>=7.4.0  # Added for future unit testing